# src/routers/task.js
const express = require('express');
const router = new express.Router();
const Task = require('../models/task');
const auth = require("../middleware/auth");

router.post('/tasks', auth, async(req, res)=>{

    const task = new Task({
        ...req.body,
        owner: req.user._id
    })

    try{
        await task.save();
        res.status(201).send(task)
    }catch(err){
        res.status(400).send(err)
    }
})

router.get('/tasks', auth, async(req, res)=>{

    const match = {}
    if(req.query.completed){
        match.completed = req.query.completed === 'true';
    }

    const sort = {}

    if(req.query.sortBy){
        sort.createdAt = req.query.sortBy == 'desc' ? -1 : 1;
    }
    
    try{
        await req.user.populate({
            path: 'tasks',
            match,
            options: {
                limit: parseInt(req.query.limit),
                skip: parseInt(req.query.skip),
                sort
            }
        }).execPopulate();
        res.send(req.user.tasks)
    }catch(err){
        res.status(500).send(err)
    }
})

router.get('/tasks/:id', auth, async(req, res)=>{

    const _id = req.params.id;

    try{
       const task= await Task.findOne({_id, owner: req.user._id});
       
       if(!task){
        return res.status(404).send();
    }
    res.send(task)
    }catch(err){
        res.status(500).send(err)
    }
})

router.patch('/tasks/:id', auth, async(req, res)=>{
    const updates = Object.keys(req.body);
    const allowedUpdates = ['description','completed']
    const isValidOperation = updates.every((update)=> allowedUpdates.includes(update))

    if(!isValidOperation){
       return res.status(404).send({error: 'Invalid updates'});
    }

    //request param - id of the task
    const _id = req.params.id;

    try{
        const task = await Task.findOne({_id, owner: req.user._id});
        
        if(!task){
            return res.status(404).send();
        }

        updates.forEach((update) => task[update] = req.body[update])

        task.save();
        res.send(task)
     }catch(err){
         res.status(500).send(err)
     }
})

router.delete('/tasks/:id', auth, async(req, res)=>{

    const _id = req.params.id;
    try{
        const task = await Task.findOneAndDelete({_id, owner: req.user._id});
        if(!task){
            return res.status(404).send();
        }
        res.send(task)
     }catch(err){
         res.status(500).send(err)
     }
})

module.exports = router;