# src/routers/user.js
const express = require('express');
const router = new express.Router();
const User = require('../models/user');
const auth = require('../middleware/auth');

router.post('/users', async (req, res)=>{

    const user = new User(req.body)

    try{
        await user.save();
        const token = await user.generateAuthToken();
        res.status(201).send({user,token})
    }catch(err){
        res.status(400).send(err)
    }
})

router.post('/users/login', async (req, res)=>{
    try{
        const user = await User.findByCredentials(req.body.email, req.body.password);
        const token = await user.generateAuthToken()
        res.send({user, token});
    }catch(err){
        console.log(err)
        res.status(400).send(err);
    }
})

router.post('/users/logout', auth, async(req,res)=>{

    try{
        req.user.tokens = req.user.tokens.filter((token)=>{
            return token.token != req.token;
        })
        await  req.user.save();

        res.send();
    }catch(e){
        res.status(500).send();
    }
})

router.post('/users/logoutAll', auth, async(req,res)=>{

    try{
        req.user.tokens = [];
        await  req.user.save();
        res.send();
    }catch(e){
        res.status(500).send();
    }
})

router.get('/users/me', auth, async (req, res)=>{
       res.send(req.user);
})

router.patch('/users/me', auth, async(req, res)=>{
    const updates = Object.keys(req.body);
    const allowedUpdates = ['name', 'password', 'age']
    const isValidOperation = updates.every((update)=> allowedUpdates.includes(update))

    if(!isValidOperation){
      return res.status(404).send({error: 'Invalid updates'});
    }

    try{
        updates.forEach((update) => req.user[update]= req.body[update]);

        await req.user.save();

        res.send(req.user)
     }catch(err){
         res.status(500).send(err)
     }
})

router.delete('/users/me', auth, async(req, res)=>{

    try{
        await req.user.remove();
        res.send(req.user)
     }catch(err){
         res.status(500).send(err)
     }
})

module.exports= router;